style: model dir structure tidy
